requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
pillow==11.0.0
jinja2==3.1.4
python-multipart==0.0.19
transformers==4.47.1
torch==2.5.1
pytest==8.3.4
httpx==0.28.1
